do not try to build mulitarch
on windows the default installation apparently still consists of i386 and x86_64. If 64 bit R is loaded, only the 64 bit RGtk2 libraries will be installed and only these will work.
Due to this installation of RQDA will fail since RGtk2 is unable to load the 32 bit Gtk2 files and the multiarch installation will fail. This failure is expected and completly irrelevant to us (you wont run 32 bit builds anyway), but will cause the entire build process to fail. Therefore we disable multiarch builds